Create README.md
I develop applications using cutting-edge technologies such as Angular, Ionic, React, while still comfortable implementing web sites and applications with pure JavaScript HTML and PHP.

I've learned that writing readable and reusable code that works efficiently is the quickest way to a robust system. In my continuous pursuit of knowledge and development, I've come to understand in better detail the ins and outs of the MVC, SOLID and DRY schools of thought.